// src/generated/ipc/models/ControlDebugSessionRequestRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Aone IDE Tauri IPC
 * Documentation paths represent Tauri invoke commands, not network HTTP endpoints. Success payloads are serialized JSON values. Backend failures reject the invocation with a serialized error string.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ControlDebugSessionRequestRequest
 */
export interface ControlDebugSessionRequestRequest {
    /**
     * 
     * @type {string}
     * @memberof ControlDebugSessionRequestRequest
     */
    runId: string;
    /**
     * 
     * @type {string}
     * @memberof ControlDebugSessionRequestRequest
     */
    debugSessionId: string;
    /**
     * 
     * @type {ControlDebugSessionRequestRequestActionEnum}
     * @memberof ControlDebugSessionRequestRequest
     */
    action: ControlDebugSessionRequestRequestActionEnum;
    /**
     * 
     * @type {number}
     * @memberof ControlDebugSessionRequestRequest
     */
    expectedSequence: number;
    /**
     * 
     * @type {number}
     * @memberof ControlDebugSessionRequestRequest
     */
    expectedControlEpoch: number;
}

/**
* @export
* @enum {string}
*/
export enum ControlDebugSessionRequestRequestActionEnum {
    Pause = 'pause',
    Resume = 'resume',
    StepInto = 'stepInto',
    StepOver = 'stepOver',
    Stop = 'stop'
}


/**
 * Check if a given object implements the ControlDebugSessionRequestRequest interface.
 */
export function instanceOfControlDebugSessionRequestRequest(value: object): value is ControlDebugSessionRequestRequest {
    if (!('runId' in value) || value['runId'] === undefined) return false;
    if (!('debugSessionId' in value) || value['debugSessionId'] === undefined) return false;
    if (!('action' in value) || value['action'] === undefined) return false;
    if (!('expectedSequence' in value) || value['expectedSequence'] === undefined) return false;
    if (!('expectedControlEpoch' in value) || value['expectedControlEpoch'] === undefined) return false;
    return true;
}

export function ControlDebugSessionRequestRequestFromJSON(json: any): ControlDebugSessionRequestRequest {
    return ControlDebugSessionRequestRequestFromJSONTyped(json, false);
}

export function ControlDebugSessionRequestRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ControlDebugSessionRequestRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'runId': json['runId'],
        'debugSessionId': json['debugSessionId'],
        'action': json['action'],
        'expectedSequence': json['expectedSequence'],
        'expectedControlEpoch': json['expectedControlEpoch'],
    };
}

export function ControlDebugSessionRequestRequestToJSON(json: any): ControlDebugSessionRequestRequest {
    return ControlDebugSessionRequestRequestToJSONTyped(json, false);
}

export function ControlDebugSessionRequestRequestToJSONTyped(value?: ControlDebugSessionRequestRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'runId': value['runId'],
        'debugSessionId': value['debugSessionId'],
        'action': value['action'],
        'expectedSequence': value['expectedSequence'],
        'expectedControlEpoch': value['expectedControlEpoch'],
    };
}