// src/generated/ipc/models/StopRunRequestRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Aone IDE Tauri IPC
 * Documentation paths represent Tauri invoke commands, not network HTTP endpoints. Success payloads are serialized JSON values. Backend failures reject the invocation with a serialized error string.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface StopRunRequestRequest
 */
export interface StopRunRequestRequest {
    /**
     * 
     * @type {string}
     * @memberof StopRunRequestRequest
     */
    runId: string;
}

/**
 * Check if a given object implements the StopRunRequestRequest interface.
 */
export function instanceOfStopRunRequestRequest(value: object): value is StopRunRequestRequest {
    if (!('runId' in value) || value['runId'] === undefined) return false;
    return true;
}

export function StopRunRequestRequestFromJSON(json: any): StopRunRequestRequest {
    return StopRunRequestRequestFromJSONTyped(json, false);
}

export function StopRunRequestRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): StopRunRequestRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'runId': json['runId'],
    };
}

export function StopRunRequestRequestToJSON(json: any): StopRunRequestRequest {
    return StopRunRequestRequestToJSONTyped(json, false);
}

export function StopRunRequestRequestToJSONTyped(value?: StopRunRequestRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'runId': value['runId'],
    };
}